fed_backend/pyproject.toml
[tool.poetry]
name = "fed-my"
version = "0.1.0"
description = ""
authors = ["yoonki <yoonkeee@gmail.com>"]
readme = "README.md"
packages = [{include = "fed_my"}]

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.0.0"
black = "^22.10.0"
selenium = "^4.7.2"
requests = "^2.28.1"
bs4 = "^0.0.1"
textblob = "^0.17.1"
nltk = "^3.8"
corpora = "^1.0"
openai = "^0.25.0"
spacy = "^3.4.4"
pymysql = "^1.0.2"
django-environ = "^0.9.0"
fastapi = "^0.89.0"
uvicorn = "^0.20.0"
sqlalchemy = "^1.4.46"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
